issue #308: 開発 host の Debian (trixie/glibc 2.41) 移行に伴う test/dist の distro fix
- tests/scripts/real-coreutils.sh: 固定 lib リストに libcap.so.2 を追加
  (Debian の coreutils ls が libcap.so.2 に依存。Ubuntu の ls では不要だった)
- tests/binaries/src/sys_sa_restart64.c / sys_sigchld64.c: wait4 の status を
  long → int に修正 (kernel ABI で status は 32-bit int*。gcc-14 で顕在化)
- dist/build-demo-bundle.sh: 同梱 glibc 表記を 2.39 → 2.41 (Debian 13 trixie)
- dist/NOTICE.txt / THIRD-PARTY-LICENSES.md: 配布 rootfs の派生元を Ubuntu →
  Debian に更新 (source archive を deb.debian.org に変更)、版数表記も更新
  (Emulin 0.5.0 / Temurin JDK 25 / Debian 13 trixie)

+++ tests/binaries/src/sys_sa_restart64.c
@@ -50,7 +50,7 @@ void _start(void) {
 
     /* 親: wait4(-1) で子を待つ。SA_RESTART のおかげで EINTR にならず
        handler 復帰後に再開し、最終的に子の終了を取得できる。 */
-    long status = 0xdeadbeef;
+    int status = 0xdeadbeef;
     long ret = sys_wait4(-1, &status, 0, 0);
 
     put("ret>0=");
